src/third_party/blink/web_tests/shadow-dom/declarative-shadow-dom-errors-expected.txt

CONSOLE ERROR: line 17: This element does not support attachShadow
CONSOLE ERROR: line 37: attachShadow() is disabled by disabledFeatures static field.
CONSOLE ERROR: line 50: Shadow root cannot be created on a host which already hosts a shadow tree.
PASS successfullyParsed is true

TEST COMPLETE
